Here's the other way to do UI -- handing down a Beard reference and an identifier for what DOM element the new UI element is supposed to add itself to.
I think I'm going to like this better.  It's going to turn out to be much more flexible when more complicated UI controllers start cropping up and wanting to wangle tons of DOM elements from a single java class or function.

Another option I was briefly toying with was to add new elements to a '#tmp' DOM element (just so that they'd be gettable again by jquery elsewhere) and returning the new element's ID instead of demanding an ID for the parent DOM object as an argument.  Seems a little messy.