Rise can now run standalone and as an applet.
The JavaFX runtime jar now gets shipped as a library, but it is not included in the same jar as the rest of Rise, which means the rise.jar file still contains only what the applet needs, and adding the JavaFX jar to the classpath is all that's needed to run standable.

The exact command for starting Rise standalone is "java -cp rise.jar:lib/javafxrt/* us.exultant.beard.LaunchStandalone us.exultant.rise.exe.Rise"... this is a little verbose, but can be knocked down pretty quick without much work.