Write mouse reactions that change borders using the full beard bus infrastructure.
This is a fairly contrived use, since being a pure UI thing there's no strong reason not to just leave handlers in the javascript.  However, this exact same mechanism could be used to kick off arbitrary work in java, up to and including spinning off more threads or whathaveyou, so that's cool.